Arquillian integration with remote WebLogic 12c using the WebLogic JSR-88 deployment protocol implementation.
What is this Arquillian container integration for?

This implementation provides the Arquillian integration with your
remote WebLogic 12c
container using the WebLogic JSR-88 deployment protocol implementation.

What are the main features of this integration?

1. As you already may know the Oracle WebLogic Server 12c is now
certified for the full
Java EE 6 platform specification, which enables higher developer
productivity with standards
based, modern APIs, including Servlet 3.0, JAX-RS 1.1, Java Server
Faces 2.1, EJB 3.1, Context
and Dependency Injection for Java, and many others. And now it can be
used with Arquallian.

2. This integration is based on the standard JSR-88 deployment API,
having a very simplified
configuration to work (in your arquillian.xml only adminUser,
andminPasswor with operator
roll needed to set as minimum, but have only 5 parameters all together).

3. The Aquillian container and the application server can run on
different nodes/machines,
but you need the WebLogic 12c Client as a dependency, it should be on
you classpath during
testing.

4. Works with a stand-alone server and clustered environment as well.

5. It has already been tested with ZIP Distribution, which is very good
for development purpose,
but having only one target that is the administration server as well.
But you never know, if you
would actually find a bug, please raise a ticket:-).

The WebLogic client (wlfullclient.jar) which is not available on any
well known public maven
repositories (as of the writing of this container integration). But it
is very easy to install
that dependency in your local repository to make it available on your
classpath for tests.

For more details please find the README.txt