ARQ-725 Improved the performance of the WLS container integration.
The JMX classloader is initalized on creation of a WebLogicJMXClient
instance. Also, the connection to the JMX MBean Server is also created
at object initialization. The WebLogicJMXClient is created only once,
for an instance of the WebLogicContainer class. The state of the JVM
modified by the WebLogicJMXClient is stashed for the duration of the
method invocations on this class, and restored on completion.

Modified the codesource of the WebLogicJMXLibClassLoader.
The codesource is now weblogic.jar instead of wljmxclient.jar to avoid
setting up a SecurityManager which in turn reduces performance.

The jmxLibPath property in arquillian.xml is no longer supported.