ARQ-707 Initial implementation of a WebLogic 12c integration.
This integration is based on the WLS 10.3 integration. The Maven module
and source packages have been renamed. The WebLogicCDIProcessor class
has been removed from the WLS12c implementation. The examples have been updated to use the Java EE 6 features.

Extracted common functionality from the WLS 10.3.x and 12c
container implementations into the arquillian-wls-common module.

The DeployableContainer and ContainerConfiguration implementations are
present in the arquillian-wls-remote-10.3 and arquillian-wls-remote-12.1
modules.

This organization assumes that deployment of archives and introspection
of deployments can be done in a common manner across several WLS
versions, and this is true of the 10.3.x and 12.1.1 versions as of now.