ARQ-785 Initial implementation of managed WebLogic containers.
The managed container implementation uses the startWebLogic and stopWebLogic
scripts in the WebLogic domain directory to control the WLS container.

Refactored the Container configuration classes. Introduced
ManagedContainer and RemoteContainer classes to encapsulate operations
on managed and remote WLS containers.

Refactored and improved the test suite for WebLogic containers.
Introduced tests for verifying-
 1. JAR deployment,
 2. Testable API of Arquillian-Core 1.0.2
 3. Multiple deployments in a test class

Segregated the others into tests to verify deployment of EARs and WARs,
CDI injection, Resource and EJB injection, ArquillianResource injection
of URLs and tests run at the client.

Detects if the startups and shutdown script has not completed by a
pre-determined timeout period, and terminates the script if necessary.