ARQ-1172 Initial implementation of Arquillian adapter for embedded WLS.
Rationale behind Surefire configuration values is explained in the POM.
For now, developers i.e. end-users are expected to make the
embeddable EJB SPI implementation of WLS 12c, in the Surefire classpath.

Developers may do so via a Maven artifact if it makes all the classes
available (currently not possible), or add weblogic.jar as a classpath
element to Surefire.

The JBoss and Oracle/Sun Java EE 6 API artifacts must be omitted from
the Surefire classpath for various reasons. The former for having
entries to JBoss classes in SPI files, and the latter for not having
method bodies. Use weblogic.jar or an equivalent to provide EE 6 APIs at
runtime.

Multiple test-scoped deployments are not supported since the second and
subsequent deployments prevents the container from cleaning up properly.